changed speed message to take all float values
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@274 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Geos/newWave.h
@@ -89,7 +89,7 @@ class GEM_EXTERN newWave : public GemShape
   void 		getposition(void);
   void		getTexCoords(void);
   void		setSize( int value );
-  void		setSpeed( int value );
+  void		setSpeed( float value );
 
   void copy( float vec0[3], float vec1[3]);
   void sub(float vec0[3], float vec1[3], float vec2[3]);
